+++ tests/test_audio_processor.cpp
@@ -40,7 +40,7 @@ TEST(AudioProcessor, PassThrough)
 	processor->Flush();
 
 	ASSERT_EQ(data.size(), buffer->data().size());
-	for (int i = 0; i < data.size(); i++) {
+	for (size_t i = 0; i < data.size(); i++) {
 		ASSERT_EQ(data[i], buffer->data()[i]) << "Signals differ at index " << i;
 	}
 }
@@ -57,7 +57,7 @@ TEST(AudioProcessor, StereoToMono)
 	processor->Flush();
 
 	ASSERT_EQ(data2.size(), buffer->data().size());
-	for (int i = 0; i < data2.size(); i++) {
+	for (size_t i = 0; i < data2.size(); i++) {
 		ASSERT_EQ(data2[i], buffer->data()[i]) << "Signals differ at index " << i;
 	}
 }
@@ -74,7 +74,7 @@ TEST(AudioProcessor, ResampleMono)
 	processor->Flush();
 
 	ASSERT_EQ(data2.size(), buffer->data().size());
-	for (int i = 0; i < data2.size(); i++) {
+	for (size_t i = 0; i < data2.size(); i++) {
 		ASSERT_EQ(data2[i], buffer->data()[i]) << "Signals differ at index " << i;
 	}
 }
@@ -91,7 +91,7 @@ TEST(AudioProcessor, ResampleMonoNonInteger)
 	processor->Flush();
 
 	ASSERT_EQ(data2.size(), buffer->data().size());
-	for (int i = 0; i < data2.size(); i++) {
+	for (size_t i = 0; i < data2.size(); i++) {
 		ASSERT_EQ(data2[i], buffer->data()[i]) << "Signals differ at index " << i;
 	}
 }
@@ -108,7 +108,7 @@ TEST(AudioProcessor, StereoToMonoAndResample)
 	processor->Flush();
 
 	ASSERT_EQ(data2.size(), buffer->data().size());
-	for (int i = 0; i < data2.size(); i++) {
+	for (size_t i = 0; i < data2.size(); i++) {
 		ASSERT_EQ(data2[i], buffer->data()[i]) << "Signals differ at index " << i;
 	}
 }
